Obtener los nombres que coinciden con un usuario

+++ crud/src/main/java/smartio/api/crud/models/UserScoreProjection.java
@@ -5,5 +5,5 @@ public interface UserScoreProjection {
     long getUserId();
     String getUserName();
     int getIntelligence();
-    int getScores();
+    float getScores();
 }
